Fix JP/Greenery Day holiday definition
In 2007, Greenery Day moved to May 4, and April 29 was changed to Shōwa Day in accordance with a 2005 revision of the law pertaining to public holidays. This fixes an issue where duplicate rows caused the `LOOKUPVALUE` function to fail with the error "A table of multiple values was supplied where a single value was expected" while retrieving the noliday name.